+++ README.md
@@ -1,2 +1,2 @@
 # Flexray-Data-Parser
-  python main.py C:\Users\hassa\OneDrive\Documents\flexray\gen.json C:\Users\hassa\OneDrive\Documents\flexray\20181107132300_bus_signals.json
+  python a2d2_data_reader.py C:\Users\hassa\OneDrive\Documents\flexray\gen.json C:\Users\hassa\OneDrive\Documents\flexray\20181107132300_bus_signals.json
