--- __init__.py
@@ -1,13 +1,0 @@
-# a2d2_dataset/__init__.py
-
-"""
-A2D2 Dataset Package
-
-This package provides functionality to load and process
-3D label data and FlexRay data from the A2D2 dataset.
-
-Modules:
-- data_models: Contains the data structures for 3D labels and FlexRay data.
-- parser: Contains functions to load and parse the JSON files.
-- dataset: Contains the A2D2Dataset class to handle matching and retrieving records.
-"""
