+++ utils/validators.py
@@ -1,0 +1,19 @@
+# a2d2_dataset/utils/validators.py
+
+import logging
+
+def validate_3d_labels(data):
+    """
+    Placeholder function to validate 3D labels JSON structure.
+    """
+    logger = logging.getLogger("validate_3d_labels")
+    # Implement actual validation logic here
+    pass
+
+def validate_flexray_data(data):
+    """
+    Placeholder function to validate FlexRay JSON structure.
+    """
+    logger = logging.getLogger("validate_flexray_data")
+    # Implement actual validation logic here
+    pass
